+++ src/app/app.module.ts
@@ -2,6 +2,8 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MAT_FORM_FIELD_DEFAULT_OPTIONS, MatFormFieldModule } from '@angular/material/form-field';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -9,32 +11,30 @@ import { MatRadioModule } from '@angular/material/radio';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatStepperModule } from '@angular/material/stepper';
-import { MatTabsModule } from '@angular/material/tabs';
+import { TextFieldModule } from '@angular/cdk/text-field';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule, Routes } from '@angular/router';
 import { p2tHttpService } from './Services/p2tHttpService';
 import { t2pHttpService } from './Services/t2pHttpService';
 import { AppComponent } from './app.component';
+import { CombinedComponent } from './components/components';
 import { HomeComponent } from './home/home.component';
-import { P2tComponent } from './p2t/p2t.component';
-import { T2PComponent } from './t2p/t2p.component';
 
 const routes: Routes = [
   { path: 'home', component: HomeComponent },
-  { path: 't2p', component: T2PComponent },
-  { path: 'p2t', component: P2tComponent },
   { path: '', redirectTo: '/home', pathMatch: 'full' },
 ];
 
 @NgModule({
-  declarations: [AppComponent, T2PComponent, HomeComponent, P2tComponent],
+  declarations: [AppComponent, HomeComponent, CombinedComponent],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     RouterModule.forRoot(routes),
     HttpClientModule,
-    MatTabsModule,
+    MatButtonModule,
+    MatCheckboxModule,
     MatStepperModule,
     MatSlideToggleModule,
     MatFormFieldModule,
@@ -43,6 +43,7 @@ const routes: Routes = [
     MatSelectModule,
     MatRadioModule,
     MatProgressSpinnerModule,
+    TextFieldModule,
     FormsModule,
     ReactiveFormsModule,
   ],
