+++ src/app/components/components.css
@@ -1,0 +1,126 @@
+.card {
+  border-radius: 4px;
+  border: 1px solid #ddd;
+  background-color: #fafafa;
+  height: 40px;
+  width: 200px;
+  margin: 0 8px 16px;
+  padding: 16px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  transition: all 0.2s ease-in-out;
+  line-height: 24px;
+  cursor: pointer;
+  box-shadow: 0 1px 4px rgba(0, 0, 0, 0.08);
+}
+
+.card:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 4px 17px rgba(0, 0, 0, 0.2);
+  border-color: #bbb;
+}
+
+.card-container {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 8px;
+}
+
+.download-card {
+  width: 220px;
+  height: 48px;
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.tool-selection {
+  display: flex;
+  gap: 16px;
+  margin-top: 8px;
+}
+
+.tool-card {
+  flex: 1;
+  border-radius: 4px;
+  border: 2px solid #ddd;
+  background-color: #fafafa;
+  padding: 16px;
+  cursor: pointer;
+  transition: all 0.2s ease-in-out;
+}
+
+.tool-card:hover {
+  border-color: #1976d2;
+  box-shadow: 0 2px 8px rgba(25, 118, 210, 0.15);
+}
+
+.tool-card.selected {
+  border-color: #1976d2;
+  background-color: #e3f2fd;
+}
+
+.dropzone {
+  border: 1px dashed #888;
+  height: 100px;
+  width: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  border-radius: 4px;
+  margin: 8px 0;
+  color: #666;
+}
+
+.insertModel {
+  margin-bottom: 8px;
+}
+
+#model-container {
+  width: 100%;
+  height: 30vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+#loading {
+  text-align: center;
+  margin: 16px 0;
+}
+
+#error-container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+#error-container-text {
+  color: white;
+  padding: 20px;
+  border-color: #da0000;
+  border-style: solid;
+  background-color: #f7614e;
+  border-radius: 10px;
+  display: none;
+}
+
+#error-box {
+  border-radius: 4px;
+  border: 1px solid #da0000;
+  background-color: #f7614e;
+  padding: 12px 16px;
+  margin: 8px 0;
+}
+
+#error-content {
+  color: white;
+  margin: 0;
+}
+
+.rag-toggle {
+  margin-left: 16px;
+}
