+++ src/app/home/home.component.html
@@ -8,17 +8,8 @@
 </div>
 
 <div class="content" role="main">
-  <!-- Tabs to Switch between T2P and P2T -->
-  <mat-tab-group id="tab-group">
-    <mat-tab label="T2P (Text-to-Process)" class="tabGroup">
-      <app-t2p></app-t2p>
-    </mat-tab>
-    <mat-tab label="P2T (Process-to-Text)" class="tabGroup">
-      <app-p2t></app-p2t>
-    </mat-tab>
-  </mat-tab-group>
 
-
+  <app-components></app-components>
 
   <!-- Links -->
   <div class="card-container">
@@ -71,10 +62,7 @@
         </g>
       </svg>
     </a>
-
-
   </div>
-
 
   <svg id="clouds" alt="Gray Clouds Background" xmlns="http://www.w3.org/2000/svg" width="2611.084" height="485.677"
     viewBox="0 0 2611.084 485.677">
@@ -83,11 +71,3 @@
       transform="translate(142.69 -634.312)" fill="#eee" />
   </svg>
 </div>
-
-<!-- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * -->
-<!-- * * * * * * * * * * * The content above * * * * * * * * * * * -->
-<!-- * * * * * * * * * * is only a placeholder * * * * * * * * * * -->
-<!-- * * * * * * * * * * and can be replaced. * * * * * * * * * * * -->
-<!-- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * -->
-<!-- * * * * * * * * * * End of Placeholder * * * * * * * * * * * -->
-<!-- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * -->
