+++ src/app/home/home.module.ts
@@ -1,48 +1,11 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
-import { HttpClientModule } from '@angular/common/http';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-import { MatInputModule } from '@angular/material/input';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-
-import { TextFieldModule } from '@angular/cdk/text-field';
-import {
-  MAT_FORM_FIELD_DEFAULT_OPTIONS,
-  MatFormFieldModule,
-} from '@angular/material/form-field';
-import { MatStepperModule } from '@angular/material/stepper';
-import { MatTabsModule } from '@angular/material/tabs';
-import { P2tComponent } from '../p2t/p2t.component';
-import { T2PComponent } from '../t2p/t2p.component';
-import { T2PModule } from '../t2p/t2p.module';
 import { HomeComponent } from './home.component';
 
 @NgModule({
   declarations: [HomeComponent],
-
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    HttpClientModule,
-    MatSlideToggleModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatTabsModule,
-    MatStepperModule,
-    TextFieldModule,
-    T2PComponent,
-    T2PModule,
-    P2tComponent,
-  ],
-  providers: [
-    {
-      provide: MAT_FORM_FIELD_DEFAULT_OPTIONS,
-      useValue: { appearance: 'outline' },
-    },
-  ],
-
+  exports: [HomeComponent],
+  imports: [BrowserModule],
   bootstrap: [HomeComponent],
 })
 export class HomeModule {}
