--- src/app/t2p/t2p.component.html
@@ -1,131 +1,0 @@
-<mat-vertical-stepper [linear]="false" #stepper [animationDuration]="200">
-  <!-- Step 1: Explanation + LLM-Konfiguration -->
-  <mat-step>
-    <form>
-      <ng-template matStepLabel>Explanation</ng-template>
-      <p>
-        <b>What is T2P?</b><br />
-        The text-to-process approach deals with the conversion of a text test
-        into a model within the framework of business process management (BPM)
-        and natural language processing (NLP).
-      </p>
-
-      <label>
-        <input type="checkbox" [checked]="isLLMEnabled" (change)="isLLMEnabled = $event.target.checked" />
-        Use LLM
-      </label>
-
-      <div *ngIf="isLLMEnabled">
-        <!-- LLM Provider Auswahl -->
-        <div class="mt-2">
-          <label class="d-block mb-1"><strong>LLM Provider</strong></label>
-          <mat-radio-group [(ngModel)]="selectedLLMProvider" name="llmProvider">
-            <mat-radio-button value="openai" class="mr-4">OpenAI</mat-radio-button>
-            <mat-radio-button value="gemini">Gemini</mat-radio-button>
-          </mat-radio-group>
-        </div>
-
-        <!-- API Key Input -->
-        <mat-form-field appearance="outline" class="w-100 mt-2">
-          <mat-label>API Key</mat-label>
-          <input matInput #apiKeyInput [(ngModel)]="apiKey" name="apiKey"
-            placeholder="Enter your LLM API key" />
-        </mat-form-field>
-        <button mat-button class="card" type="button" (click)="validateApiKey()" [disabled]="apiKeyChecking">
-          {{ apiKeyChecking ? 'Checking...' : 'Validate Key' }}
-        </button>
-        <p style="color: green;" *ngIf="apiKeyValid === true">✔ API Key is valid.</p>
-        <p style="color: red;" *ngIf="apiKeyValid === false">✘ Invalid API Key for {{ selectedLLMProvider }}.</p>
-
-        <!-- Prompting Strategy Auswahl -->
-        <div class="mt-2">
-          <label class="d-block mb-1"><strong>Prompting Strategy</strong></label>
-          <mat-radio-group [(ngModel)]="promptingStrategy" name="promptingStrategy">
-            <mat-radio-button value="few_shot" class="mr-4">Few-Shot</mat-radio-button>
-            <mat-radio-button value="zero_shot">Zero-Shot</mat-radio-button>
-          </mat-radio-group>
-        </div>
-      </div>
-    </form>
-    <br />
-    <button mat-button matStepperNext class="card">Next</button>
-  </mat-step>
-
-  <!-- Step 2: Text Input -->
-  <mat-step>
-    <form>
-      <ng-template matStepLabel>Insert your Text</ng-template>
-      <mat-form-field appearance="fill" [style.width.%]="50">
-        <mat-label>Enter Text to Transform</mat-label>
-        <textarea class="textfeld" [style.width.%]="100" matInput cdkTextareaAutosize cdkAutosizeMinRows="4"
-          cdkAutosizeMaxRows="100" id="text-input" #modelInput>{{ text }}</textarea>
-      </mat-form-field>
-
-      <!-- File Upload via Drag & Drop -->
-      <div class="dropzone" #dropzoneRef (drop)="onDrop($event)" (dragover)="onDragOver($event)"
-        (click)="selectFiles()">
-        <input type="file" style="display: none" #fileInputRef accept=".txt" (change)="onFileSelected($event)" />
-        <span *ngIf="isFiledDropped">Uploaded file: {{ droppedFileName }}</span>
-        <span *ngIf="!isFiledDropped">
-          Upload file via drag-and-drop or click here
-        </span>
-      </div>
-
-      <div>
-        <button mat-button matStepperPrevious class="card">Back</button>
-        <button type="button" mat-button matStepperNext class="card">
-          Next
-        </button>
-      </div>
-    </form>
-  </mat-step>
-
-  <!-- Step 3: Select Model -->
-  <mat-step>
-    <form>
-      <ng-template matStepLabel>Choose Your Model</ng-template>
-      <mat-radio-group aria-label="Select an option" (click)="onSelectedDiagram($event)">
-        <mat-radio-button value="bpmn">BPMN</mat-radio-button>
-        <mat-radio-button value="petri-net">Petri-Net</mat-radio-button>
-      </mat-radio-group>
-
-      <div>
-        <button type="button" mat-button matStepperPrevious class="card">Back</button>
-        <button type="button" mat-button matStepperNext class="card" (click)="generateProcess(modelInput.value)">
-          Generate
-        </button>
-      </div>
-    </form>
-  </mat-step>
-
-  <!-- Step 4: Output -->
-  <mat-step>
-    <form>
-      <ng-template matStepLabel>Process (Done)</ng-template>
-      <p>You are now done.</p>
-      <div>
-        <button mat-button matStepperPrevious class="card">Back</button>
-      </div>
-    </form>
-
-    <div id="loading">
-      <mat-spinner [diameter]="100" *ngIf="spinnerService.visibility | async">
-      </mat-spinner>
-    </div>
-  </mat-step>
-</mat-vertical-stepper>
-
-<!-- Output -->
-<div>
-  <label><b>Your Input:</b> {{ textResult }}</label>
-</div>
-<button (click)="onDownloadImage()" class="card">Download Process as .png</button>
-<button (click)="onDownloadText()" class="card">Download Process as File</button>
-
-<!-- Error Handling -->
-<div id="error-container">
-  <p id="error-container-text"></p>
-</div>
-
-<!-- BPMN Container -->
-<div id="model-container"></div>
